requirements.txt
openai
flask
mysql-connector-python
python-dotenv